Chapter 8 - added texture atlases (different textures stored on one texture), particle system (missle fire and smoke), animations (explosions and eagle roll), and shaders (bloom effect).
Issues - the background texture JungleTooBig was to large for my video
card.  The old Desert texture was used in its place renamed to Jungle.
Shaders performed very poorly on my video card.  The PostEffect and
BloomEffect classes were left in, but their use was commented out in
World::draw().  Also, the mSceneTexture.create() which is used w/ shaders
was commented out in World's ctor, because it is causing an access
violation (crash) when transitioning from the GameOverState to the
MenuState.