Chapter 9 - played different music themes (menus and during gameplay), added sound effects (button presses and gameplay events such as an explosion), used positional sound effects in the 2D world to convey feeling of spatial sound.  Included all necessary ogg, wav, and dll files.
Issues - sf::Music and sf:Sound both cause a known error.  An access
violation occurs in the destructor of sf::AudioDevice.  The work around is
to not destroy the sf::AudioDevice until the game shuts down.  There will
be a first-chance exception with the debugger, but no issues running the
game standalone.